add two implementations (sparse and dense) for UnsafeArrayData
